Add markdown-preview plugin
The plugin shows the markdown code in a web browser. An advantage of this plugin is that GitHub uses the same plugin which means that what I see locally will be exactly displayed in GitHub (Or to be accurate - almost the same).

For more info click [here](https://github.com/iamcco/markdown-preview.nvim).